rag_app/backend/requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu>=1.9.0
PyPDF2
numpy>=1.26.0
requests
python-multipart

